financial_ledger_idempotency/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
psycopg2-binary==2.9.9
requests==2.31.0
pytest==8.0.2
httpx==0.27.0
